Casts...

+++ lib/WeatherMapNode.class.php
@@ -615,7 +615,7 @@ class WeatherMapNode extends WeatherMapItem
 		// Draw the icon, if any
 		if(isset($icon_im))
 		{
-			imagecopy($node_im, $icon_im, $icon_x1, $icon_y1, 0, 0, imagesx($icon_im), imagesy($icon_im));
+			imagecopy($node_im, $icon_im, (int)$icon_x1, (int)$icon_y1, 0, 0, (int)imagesx($icon_im), (int)imagesy($icon_im));
 			imagedestroy($icon_im);
 		}
